# Improve NEON32 decoding speed